4.21*: blogilistan laajennus, step10; + autentikoinnin rikkomiin testeihin tunnistautuminen

+++ controllers/blogs_controller.js
@@ -27,22 +27,18 @@ app.post('/', async (request, response, next) => {
   try {
     const token = request.token
     if (!token) {
-      return response.status(401).json({
-        message: 'Authentication required'
-      })
+      const message = 'Authentication required'
+      return response.status(401).json({ message })
     }
-    console.log(request.token)
     const user = await User.findById(token.id)
     if (!user) {
-      throw Error('Impossible: unknown user id in authorization token!!!')
+      throw Error('Unknown user id in authorization token!!!')
     }    
     const blogData = request.body
     const blog = new Blog({ ...blogData, user: token.id })
     await blog.save()
-
     user.blogs.push(blog._id)
     await user.save()
-
     return response.status(200).json(blog.toJSON())
   } catch (error) {
     return next(error)
@@ -51,7 +47,17 @@ app.post('/', async (request, response, next) => {
 
 app.delete('/:id', async (request, response, next) => {
   try {
+    const token = request.token
+    if (!token) {
+      const message = 'Authentication required'
+      return response.status(401).json({ message })
+    }
     const id = request.params.id
+    const blog = await Blog.findById(id)
+    if (blog.user.toString() !== token.id) {
+      const message = 'DELETE allowed only for creator'
+      return response.status(401).json({ message })
+    }
     await Blog.findByIdAndDelete(id)
     return response.status(204).end()
   } catch (error) {
